rename: 공통 컴포넌트 폴더 이름 변경
