+++ src/app/(testPages)/calendar/dropdown/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import DropdownCalendar from "@/components/calendars/DropdownCalendar";
+import DropdownCalendar from "@/components/Calendar/DropdownCalendar";
 import useDateStore from "@/store/dateStore";
 
 export default function Page() {
